refactor Machine, fix namespaces
