+++ Build1/zmachine.Library/IIO.cs
@@ -1,8 +1,8 @@
-﻿namespace zmachine
+﻿namespace zmachine.Library
 {
     public interface IIO
     {
-        public string ReadLine();
+        public string? ReadLine();
         public void Write(string str);
         public void WriteLine(string str);
         public System.ConsoleKeyInfo ReadKey();
